Corrige tratamento de exceções e aprimora ApiExceptionHandler

+++ src/main/java/br/com/mh/mental_health_core/service/ConsultaService.java
@@ -1,9 +1,9 @@
 package br.com.mh.mental_health_core.service;
 
-import br.com.mh.mental_health_core.exceptions.ConsultaNotFoundException;
+import br.com.mh.mental_health_core.exceptions.MentalHealthException;
 import br.com.mh.mental_health_core.model.Consulta;
 import br.com.mh.mental_health_core.repository.ConsultaRepository;
-import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -11,16 +11,26 @@ import java.util.List;
 @Service
 public class ConsultaService {
 
-    @Autowired
-    private ConsultaRepository consultaRepository;
+    private final ConsultaRepository consultaRepository;
+
+    public ConsultaService(ConsultaRepository consultaRepository) {
+        this.consultaRepository = consultaRepository;
+    }
 
     public List<Consulta> getAllConsultas() {
-        return consultaRepository.findAll();
+        List<Consulta> consultas = consultaRepository.findAll();
+        if (consultas.isEmpty()) {
+            throw new MentalHealthException(HttpStatus.NO_CONTENT, "Nenhuma consulta encontrada.");
+        }
+        return consultas;
     }
 
     public Consulta getConsultaById(Integer id) {
         return consultaRepository.findById(id)
-                .orElseThrow(() -> new ConsultaNotFoundException("Consulta não encontrada com ID: " + id));
+                .orElseThrow(() -> new MentalHealthException(
+                        HttpStatus.NOT_FOUND, 
+                        "Consulta não encontrada com ID: " + id
+                ));
     }
 
     public Consulta saveConsulta(Consulta consulta) {
@@ -28,7 +38,12 @@ public class ConsultaService {
     }
 
     public void deleteConsulta(Integer id) {
-        consultaRepository.findById(id).orElseThrow(() -> new ConsultaNotFoundException("Consulta não encontrada com ID: " + id));
+        if (!consultaRepository.existsById(id)) {
+            throw new MentalHealthException(
+                    HttpStatus.NOT_FOUND, 
+                    "Consulta não encontrada com ID: " + id
+            );
+        }
         consultaRepository.deleteById(id);
     }
 }
