+++ src/main/java/br/com/mh/mental_health_core/service/DisponibilidadeService.java
@@ -1,9 +1,9 @@
 package br.com.mh.mental_health_core.service;
 
-import br.com.mh.mental_health_core.exceptions.DisponibilidadeNotFoundException;
+import br.com.mh.mental_health_core.exceptions.MentalHealthException;
 import br.com.mh.mental_health_core.model.Disponibilidade;
 import br.com.mh.mental_health_core.repository.DisponibilidadeRepository;
-import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -11,8 +11,11 @@ import java.util.List;
 @Service
 public class DisponibilidadeService {
 
-    @Autowired
-    private DisponibilidadeRepository disponibilidadeRepository;
+    private final DisponibilidadeRepository disponibilidadeRepository;
+
+    public DisponibilidadeService(DisponibilidadeRepository disponibilidadeRepository) {
+        this.disponibilidadeRepository = disponibilidadeRepository;
+    }
 
     public List<Disponibilidade> getAllDisponibilidades() {
         return disponibilidadeRepository.findAll();
@@ -20,7 +23,10 @@ public class DisponibilidadeService {
 
     public Disponibilidade getDisponibilidadeById(Integer id) {
         return disponibilidadeRepository.findById(id)
-                .orElseThrow(() -> new DisponibilidadeNotFoundException("Disponibilidade não encontrada com ID: " + id));
+                .orElseThrow(() -> new MentalHealthException(
+                        HttpStatus.NOT_FOUND, 
+                        "Disponibilidade não encontrada com ID: " + id
+                ));
     }
 
     public Disponibilidade saveDisponibilidade(Disponibilidade disponibilidade) {
@@ -28,7 +34,12 @@ public class DisponibilidadeService {
     }
 
     public void deleteDisponibilidade(Integer id) {
-        disponibilidadeRepository.findById(id).orElseThrow(() -> new DisponibilidadeNotFoundException("Disponibilidade não encontrada com ID: " + id));
+        if (!disponibilidadeRepository.existsById(id)) {
+            throw new MentalHealthException(
+                    HttpStatus.NOT_FOUND, 
+                    "Disponibilidade não encontrada com ID: " + id
+            );
+        }
         disponibilidadeRepository.deleteById(id);
     }
 }
