+++ src/main/java/br/com/mh/mental_health_core/service/PacienteService.java
@@ -1,9 +1,9 @@
 package br.com.mh.mental_health_core.service;
 
-import br.com.mh.mental_health_core.exceptions.PacienteNotFoundException;
+import br.com.mh.mental_health_core.exceptions.MentalHealthException;
 import br.com.mh.mental_health_core.model.Paciente;
 import br.com.mh.mental_health_core.repository.PacienteRepository;
-import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -11,8 +11,11 @@ import java.util.List;
 @Service
 public class PacienteService {
 
-    @Autowired
-    private PacienteRepository pacienteRepository;
+    private final PacienteRepository pacienteRepository;
+
+    public PacienteService(PacienteRepository pacienteRepository) {
+        this.pacienteRepository = pacienteRepository;
+    }
 
     public List<Paciente> getAllPacientes() {
         return pacienteRepository.findAll();
@@ -20,7 +23,10 @@ public class PacienteService {
 
     public Paciente getPacienteById(Integer id) {
         return pacienteRepository.findById(id)
-                .orElseThrow(() -> new PacienteNotFoundException("Paciente não encontrado com ID: " + id));
+                .orElseThrow(() -> new MentalHealthException(
+                        HttpStatus.NOT_FOUND, 
+                        "Paciente não encontrado com ID: " + id
+                ));
     }
 
     public Paciente savePaciente(Paciente paciente) {
@@ -28,7 +34,12 @@ public class PacienteService {
     }
 
     public void deletePaciente(Integer id) {
-        pacienteRepository.findById(id).orElseThrow(() -> new PacienteNotFoundException("Paciente não encontrado com ID: " + id));
+        if (!pacienteRepository.existsById(id)) {
+            throw new MentalHealthException(
+                    HttpStatus.NOT_FOUND, 
+                    "Paciente não encontrado com ID: " + id
+            );
+        }
         pacienteRepository.deleteById(id);
     }
 }
