+++ src/main/java/br/com/mh/mental_health_core/service/PsicologoService.java
@@ -1,9 +1,9 @@
 package br.com.mh.mental_health_core.service;
 
-import br.com.mh.mental_health_core.exceptions.PsicologoNotFoundException;
+import br.com.mh.mental_health_core.exceptions.MentalHealthException;
 import br.com.mh.mental_health_core.model.Psicologo;
 import br.com.mh.mental_health_core.repository.PsicologoRepository;
-import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -11,8 +11,11 @@ import java.util.List;
 @Service
 public class PsicologoService {
 
-    @Autowired
-    private PsicologoRepository psicologoRepository;
+    private final PsicologoRepository psicologoRepository;
+
+    public PsicologoService(PsicologoRepository psicologoRepository) {
+        this.psicologoRepository = psicologoRepository;
+    }
 
     public List<Psicologo> getAllPsicologos() {
         return psicologoRepository.findAll();
@@ -20,7 +23,10 @@ public class PsicologoService {
 
     public Psicologo getPsicologoById(Integer id) {
         return psicologoRepository.findById(id)
-                .orElseThrow(() -> new PsicologoNotFoundException("Psicólogo não encontrado com ID: " + id));
+                .orElseThrow(() -> new MentalHealthException(
+                        HttpStatus.NOT_FOUND, 
+                        "Psicólogo não encontrado com ID: " + id
+                ));
     }
 
     public Psicologo savePsicologo(Psicologo psicologo) {
@@ -28,7 +34,12 @@ public class PsicologoService {
     }
 
     public void deletePsicologo(Integer id) {
-        psicologoRepository.findById(id).orElseThrow(() -> new PsicologoNotFoundException("Psicólogo não encontrado com ID: " + id));
+        if (!psicologoRepository.existsById(id)) {
+            throw new MentalHealthException(
+                    HttpStatus.NOT_FOUND, 
+                    "Psicólogo não encontrado com ID: " + id
+            );
+        }
         psicologoRepository.deleteById(id);
     }
 }
